sentry - auth middleware: - Connect to `Redis` and pass the connection to `Application` - Move cors middleware module to it's own file - Implement auth middleware - Change `sentry::Session.ip` to `Option`
