+++ sentry/src/lib.rs
@@ -1,14 +1,19 @@
 #![deny(clippy::all)]
 #![deny(rust_2018_idioms)]
 
+use crate::middleware::auth;
 use crate::middleware::cors::{cors, CorsResult};
 use hyper::service::{make_service_fn, service_fn};
 use hyper::{Body, Error, Request, Response, Server, StatusCode};
 use primitives::adapter::Adapter;
 use primitives::Config;
+use redis::aio::SharedConnection;
 use slog::{error, info, Logger};
 
-pub mod middleware;
+pub mod middleware {
+    pub mod auth;
+    pub mod cors;
+}
 pub mod routes {
     pub mod channel;
 }
@@ -20,32 +25,44 @@ pub mod event_reducer;
 pub struct Application<A: Adapter> {
     adapter: A,
     logger: slog::Logger,
+    redis: SharedConnection,
     _clustered: bool,
     port: u16,
     config: Config,
 }
 
 impl<A: Adapter + 'static> Application<A> {
-    pub fn new(adapter: A, config: Config, logger: Logger, clustered: bool, port: u16) -> Self {
+    pub fn new(
+        adapter: A,
+        config: Config,
+        logger: Logger,
+        redis: SharedConnection,
+        clustered: bool,
+        port: u16,
+    ) -> Self {
         Self {
             adapter,
             config,
             logger,
+            redis,
             _clustered: clustered,
             port,
         }
     }
 
+    /// Starts the `hyper` `Server`.
     pub async fn run(&self) {
         let addr = ([127, 0, 0, 1], self.port).into();
         info!(&self.logger, "Listening on port {}!", self.port);
 
         let make_service = make_service_fn(move |_| {
             let adapter_config = (self.adapter.clone(), self.config.clone());
+            let redis = self.redis.clone();
             async move {
                 Ok::<_, Error>(service_fn(move |req| {
                     let adapter_config = adapter_config.clone();
-                    async move { Ok::<_, Error>(handle_routing(req, adapter_config.0).await) }
+                    let redis = redis.clone();
+                    async move { Ok::<_, Error>(handle_routing(req, adapter_config.0, redis).await) }
                 }))
             }
         });
@@ -73,7 +90,11 @@ where
     }
 }
 
-async fn handle_routing(req: Request<Body>, adapter: impl Adapter) -> Response<Body> {
+async fn handle_routing(
+    req: Request<Body>,
+    adapter: impl Adapter,
+    redis: SharedConnection,
+) -> Response<Body> {
     let headers = match cors(&req) {
         CorsResult::Simple(headers) => headers,
         // if we have a Preflight, just return the response directly
@@ -81,19 +102,30 @@ async fn handle_routing(req: Request<Body>, adapter: impl Adapter) -> Response<B
         CorsResult::None => Default::default(),
     };
 
+    // otherwise problems with `.await` occurs about `Sync` not required for `Adapter`.
+    let auth_connections = (adapter.clone(), redis.clone());
+    let req = match auth::for_request(req, auth_connections.0, auth_connections.1).await {
+        Ok(req) => req,
+        Err(response_error) => return map_response_error(response_error),
+    };
+
     let mut response = if req.uri().path().starts_with("/channel") {
         crate::routes::channel::handle_channel_routes(req, adapter).await
     } else {
         Err(ResponseError::NotFound)
     }
-    .unwrap_or_else(|response_err| match response_err {
-        ResponseError::NotFound => not_found(),
-        ResponseError::BadRequest(error) => bad_request(error),
-    });
+    .unwrap_or_else(map_response_error);
 
-    // extend the headers with the inital headers we have from CORS (if there are some)
+    // extend the headers with the initial headers we have from CORS (if there are some)
     response.headers_mut().extend(headers);
     response
+}
+
+fn map_response_error(error: ResponseError) -> Response<Body> {
+    match error {
+        ResponseError::NotFound => not_found(),
+        ResponseError::BadRequest(error) => bad_request(error),
+    }
 }
 
 pub fn not_found() -> Response<Body> {
@@ -116,5 +148,5 @@ pub fn bad_request(error: Box<dyn std::error::Error>) -> Response<Body> {
 pub struct Session {
     pub era: i64,
     pub uid: String,
-    pub ip: String,
+    pub ip: Option<String>,
 }
